requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
PyJWT==2.5.0
requests==2.28.1
SQLAlchemy==1.4.46
sukiyou-sql==1.4.1
pymysql==1.0.2
redis==4.5.1